Simulate Population Harvest

This is a IBM to generate a population and catch from the population

simulate_population_harvest(
  Linf,
  Linf_sd,
  M,
  Lorenzen,
  F,
  mincat,
  catsd,
  maxcat,
  maxcatsd,
  L0,
  L0_sd,
  k,
  k_sd,
  Amax,
  age_max,
  N
)

Arguments

Linf

Von Bertalanffy theoretical asymptotic length (cm)

Linf_sd

Population standard deviation of asymptotic length (cm)

M

Instantaneous natural mortality rate (yr-1)

Lorenzen

TRUE / FALSE specifying whether natural mortality is a function of individual length following Lorenzen (Lorenzen, 2000; Lorenzen, 2005)

F

Apical (fully selected) instantaneous fishing mortality rate (yr-1)

mincat

Minimum length at 50% fishery selectivity (cm)

catsd

Slope of the ascending region of selectivity at length (cm), see details

maxcat

Maximum length at 50% fishery selectivity (cm)

maxcatsd

Slope of the descending region of selectivity at length (cm), see details

L0

Von Bertalanffy length at age 0 (cm)

L0_sd

Population standard deviation of length at age 0 (cm)

k

Von Bertalanffy growth coefficient

k_sd

Population standard deviation of Von Bertalanffy growth coefficient

Amax

Maximum longevity (years)

age_max

An arbitrary age selected to represent “old” fish (years)

N

The number of age 0 fish in each simulated cohort, typical value =100,000